Revised roles and removed owner field
The roles were revised to make the initial creator of a league the
administrator. There can now be multiple administrators(owners) of
a league. Administrators have all permissions allowed as well now.

The owner field was removed from the documents in the leagues
collection. Newly created leagues are initialized with the user as an
administrator for the league.